requirements.txt
psycopg2-binary
# Add your web framework here, for example:
# fastapi
# uvicorn
#fastapi
#uvicorn[standard]
#sqlalchemy[asyncio]
#asyncpg
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
phonenumbers
passlib[bcrypt]
python-jose[cryptography]
httpx
aiobotocore
python-magic
slowapi
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
phonenumbers
passlib[bcrypt]
python-jose[cryptography]
httpx
aiobotocore
python-magic
slowapi

fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
pydantic-settings
phonenumbers
passlib[bcrypt]
python-jose[cryptography]
httpx
aiobotocore
python-magic
slowapi